requirements.txt
grad-cam==1.5.0
monai==1.3.1
nibabel==5.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
optuna==3.6.1
panda==0.3.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
pytorch-crf==0.7.2
pytz==2024.1
pyyaml==6.0.1
requests==2.31.0
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
segmentation-models-pytorch==0.3.3
sentry-sdk==2.3.1
setproctitle==1.3.3
sympy==1.12
tenacity==8.3.0
threadpoolctl==3.5.0
tifffile==2024.5.10
timm==0.9.2
torchmetrics==1.4.0.post0
tqdm==4.66.4
triton==2.0.0
ttach==0.0.3
typing-extensions==4.11.0
tzdata==2024.1
unzip==1.0.0
urllib3==2.2.1
wandb==0.17.0

